Moved login authentication from login controller to a new authLogin controller
